+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added `leadsTo` temporal operator (#1932)
+
 ### Changed
 
 - TLC backend now ensures Apalache distribution is available locally before running TLC
@@ -18,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 ### Fixed
 
+- Fixed a bug in fairness operators where the effect system would report an error (#1932)
 - Fixed TLC backend failing when multiple instances run in parallel (#1949)
 - Fixed an issue where the rust evaluator failed to launch on Windows (#1945) 
 - Fixed an import issue affecting yargs on node 27 (#1926)